+++ bashrc-aliases.sh
@@ -1,6 +1,27 @@
 #!/usr/bin/env bash
-# Microsoft/Rush/SharePoint specific aliases and functions
-# Only source this on machines working with Microsoft projects
+# Shell aliases and functions
+# General-purpose and project-specific shortcuts
+
+# Display and navigation
+alias ls='ls --show-control-chars -F --color $*'
+alias pwd='cd'
+alias clear='cls'
+alias vi='vim $*'
+alias ci='code-insiders $*'
+
+# Git aliases
+alias gl='git log --oneline --all --graph --decorate  $*'
+alias gp='git pull origin --prune'
+alias gps='git push origin $*'
+alias gcm='git checkout main'
+alias ga='git add -A'
+alias gs='git status'
+alias gmm='git merge main'
+alias gc='git gc --prune=now && git remote prune origin'
+alias gf='git fetch origin'
+
+# Git config - pretty log format
+git config --global alias.lg "log --color --graph --pretty=format:'%Cred%h%Creset -%C(yellow)%d%Creset %s %Cgreen(%cr) %C(bold blue)<%an>%Creset' --abbrev-commit"
 
 # Rush build functions
 rb() {
@@ -40,7 +61,7 @@ rbaia() {
   rush install && rush build -t tag:chatodsp-pages -t sp-pages && rush dev-deploy -t tag:chatodsp-pages -t sp-pages
 }
 
-# Git function for Microsoft user branches
+# Git function for user branches
 gcb() {
   git checkout -b "user/liaye/$@"
 }